python/src/auth/requirements.txt
astroid==3.3.9
blinker==1.9.0
click==8.1.8
dill==0.3.9
Flask==3.1.0
Flask-MySQLdb==2.0.0
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
mysqlclient==2.2.7
parso==0.8.4
platformdirs==4.3.7
PyJWT==2.10.1
pylint==3.3.6
tomlkit==0.13.2
Werkzeug==3.1.3
